Cleanup direct query resolution
