+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/FulltextSearchSvcImpl.java
@@ -27,6 +27,7 @@ import ca.uhn.fhir.jpa.dao.search.ExtendedLuceneClauseBuilder;
 import ca.uhn.fhir.jpa.dao.search.ExtendedLuceneIndexExtractor;
 import ca.uhn.fhir.jpa.dao.search.ExtendedLuceneSearchBuilder;
 import ca.uhn.fhir.jpa.dao.search.LastNOperation;
+import ca.uhn.fhir.jpa.dao.search.ExtendedLuceneResourceProjection;
 import ca.uhn.fhir.jpa.model.entity.ResourceTable;
 import ca.uhn.fhir.jpa.model.search.ExtendedLuceneIndexData;
 import ca.uhn.fhir.jpa.search.autocomplete.ValueSetAutocompleteOptions;
@@ -35,7 +36,6 @@ import ca.uhn.fhir.jpa.searchparam.SearchParameterMap;
 import ca.uhn.fhir.jpa.searchparam.extractor.ISearchParamExtractor;
 import ca.uhn.fhir.jpa.searchparam.extractor.ResourceIndexedSearchParams;
 import ca.uhn.fhir.model.api.IQueryParameterType;
-import ca.uhn.fhir.model.primitive.IdDt;
 import ca.uhn.fhir.parser.IParser;
 import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.server.RequestDetails;
@@ -271,40 +271,13 @@ public class FulltextSearchSvcImpl implements IFulltextSearchSvc {
 		return convertLongsToResourcePersistentIds(pidList);
 	}
 
-	public static class ResourceProjection {
-		final long myPid;
-		final String myForcedId;
-		final String myResourceString;
-
-		public ResourceProjection(long thePid, String theForcedId, String theResourceString) {
-			myPid = thePid;
-			myForcedId = theForcedId;
-			myResourceString = theResourceString;
-		}
-
-		public IBaseResource toResource(IParser theParser) {
-			// fixme test
-			IBaseResource result = theParser.parseResource(myResourceString);
-
-			IdDt id;
-			if (myForcedId != null) {
-				id = new IdDt(myForcedId);
-			} else {
-				id = new IdDt(myPid);
-			}
-			result.setId(id);
-
-			return result;
-		}
-	}
-
 	@Override
 	public List<IBaseResource> getResources(Collection<Long> thePids) {
 		SearchSession session = getSearchSession();
-		List<ResourceProjection> rawResourceDataList = session.search(ResourceTable.class)
+		List<ExtendedLuceneResourceProjection> rawResourceDataList = session.search(ResourceTable.class)
 			.select(
 				f -> f.composite(
-					(pid, forcedId, resource)-> new ResourceProjection(pid, forcedId, resource),
+					(pid, forcedId, resource)-> new ExtendedLuceneResourceProjection(pid, forcedId, resource),
 					f.field("myId", Long.class),
 					f.field("myForcedId", String.class),
 					f.field("myRawResource", String.class))
